requirements.txt
discord.py>=2.3,<3.0
anthropic>=0.40,<1.0
